[feature] Allow configurations and severities to be passed into plugins.
We need to be able to pass configurations into plugins to decouple them from specific use cases.
Adding severities means that we can track what is a build-breaker or just a courtesy warning, as well as being able to easily silence rules without deleting them.

BREAKING CHANGE:

A list of plugins is no longer an accepted format. Configs should be structured as follows:
```
{
  'polish-plugin-banned-selectors`: [2, { selectors: ['.bad'] }]
}
```
The first argument for each plugin must be a severity level: 0 (ignored), 1 (warning), or 2 (error).
The second argument must be an object, with keys that match options that a plugin author has specified.